+++ packages/uba/modules/publicRegistration.js
@@ -1,5 +1,8 @@
 /* global ubs_settings uba_user uba_otp */
 const http = require('http')
+const UB = require('@unitybase/ub')
+const Session = UB.Session
+const App = UB.App
 
 // eslint-disable-next-line no-useless-escape
 const EMAIL_VALIDATION_RE = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/